backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
